test: move more BrowserWindow specs to main runner
