+++ spec/api-browser-window-spec.js
@@ -113,39 +113,6 @@ describe('BrowserWindow module', () => {
 
   afterEach(closeTheWindow)
 
-  describe('BrowserWindow.showInactive()', () => {
-    it('should not focus on window', () => {
-      w.showInactive()
-      expect(w.isFocused()).to.be.false()
-    })
-  })
-
-  describe('BrowserWindow.focus()', () => {
-    it('does not make the window become visible', () => {
-      expect(w.isVisible()).to.be.false()
-      w.focus()
-      expect(w.isVisible()).to.be.false()
-    })
-  })
-
-  describe('BrowserWindow.blur()', () => {
-    it('removes focus from window', () => {
-      w.blur()
-      expect(w.isFocused()).to.be.false()
-    })
-  })
-
-  describe('BrowserWindow.getFocusedWindow()', () => {
-    it('returns the opener window when dev tools window is focused', (done) => {
-      w.show()
-      w.webContents.once('devtools-focused', () => {
-        expect(BrowserWindow.getFocusedWindow()).to.equal(w)
-        done()
-      })
-      w.webContents.openDevTools({ mode: 'undocked' })
-    })
-  })
-
   describe('BrowserWindow autoHideMenuBar property', () => {
     it('has an autoHideMenuBar property', () => {
       expect(w).to.have.a.property('autoHideMenuBar')
@@ -153,48 +120,6 @@ describe('BrowserWindow module', () => {
       // TODO(codebytere): remove when propertyification is complete
       expect(w.setAutoHideMenuBar).to.be.a('function')
       expect(w.isMenuBarAutoHide).to.be.a('function')
-    })
-  })
-
-  describe('BrowserWindow.moveTop()', () => {
-    it('should not steal focus', async () => {
-      const posDelta = 50
-      const wShownInactive = emittedOnce(w, 'show')
-      w.showInactive()
-      await wShownInactive
-      expect(w.isFocused()).to.be.false()
-
-      const otherWindow = new BrowserWindow({ show: false, title: 'otherWindow' })
-      const otherWindowShown = emittedOnce(otherWindow, 'show')
-      otherWindow.loadURL('data:text/html,<html><body background-color: rgba(255,255,255,0)></body></html>')
-      otherWindow.show()
-      await otherWindowShown
-      expect(otherWindow.isFocused()).to.be.true()
-
-      w.moveTop()
-      const wPos = w.getPosition()
-      const wMoving = emittedOnce(w, 'move')
-      w.setPosition(wPos[0] + posDelta, wPos[1] + posDelta)
-      await wMoving
-      expect(w.isFocused()).to.be.false()
-      expect(otherWindow.isFocused()).to.be.true()
-
-      const wFocused = emittedOnce(w, 'focus')
-      w.focus()
-      await wFocused
-      expect(w.isFocused()).to.be.true()
-
-      otherWindow.moveTop()
-      const otherWindowPos = otherWindow.getPosition()
-      const otherWindowMoving = emittedOnce(otherWindow, 'move')
-      otherWindow.setPosition(otherWindowPos[0] + posDelta, otherWindowPos[1] + posDelta)
-      await otherWindowMoving
-      expect(otherWindow.isFocused()).to.be.false()
-      expect(w.isFocused()).to.be.true()
-
-      await closeWindow(otherWindow, { assertSingleWindow: false }).then(() => {
-        expect(BrowserWindow.getAllWindows()).to.have.lengthOf(2) // Test window + w
-      })
     })
   })
 
@@ -228,234 +153,6 @@ describe('BrowserWindow module', () => {
       // Check the 25th byte in the PNG.
       // Values can be 0,2,3,4, or 6. We want 6, which is RGB + Alpha
       expect(imgBuffer[25]).to.equal(6)
-    })
-  })
-
-  describe('BrowserWindow.setBounds(bounds[, animate])', () => {
-    it('sets the window bounds with full bounds', () => {
-      const fullBounds = { x: 440, y: 225, width: 500, height: 400 }
-      w.setBounds(fullBounds)
-
-      expectBoundsEqual(w.getBounds(), fullBounds)
-    })
-
-    it('sets the window bounds with partial bounds', () => {
-      const fullBounds = { x: 440, y: 225, width: 500, height: 400 }
-      w.setBounds(fullBounds)
-
-      const boundsUpdate = { width: 200 }
-      w.setBounds(boundsUpdate)
-
-      const expectedBounds = Object.assign(fullBounds, boundsUpdate)
-      expectBoundsEqual(w.getBounds(), expectedBounds)
-    })
-  })
-
-  describe('BrowserWindow.setSize(width, height)', () => {
-    it('sets the window size', async () => {
-      const size = [300, 400]
-
-      const resized = emittedOnce(w, 'resize')
-      w.setSize(size[0], size[1])
-      await resized
-
-      expectBoundsEqual(w.getSize(), size)
-    })
-  })
-
-  describe('BrowserWindow.setMinimum/MaximumSize(width, height)', () => {
-    it('sets the maximum and minimum size of the window', () => {
-      expect(w.getMinimumSize()).to.deep.equal([0, 0])
-      expect(w.getMaximumSize()).to.deep.equal([0, 0])
-
-      w.setMinimumSize(100, 100)
-      expectBoundsEqual(w.getMinimumSize(), [100, 100])
-      expectBoundsEqual(w.getMaximumSize(), [0, 0])
-
-      w.setMaximumSize(900, 600)
-      expectBoundsEqual(w.getMinimumSize(), [100, 100])
-      expectBoundsEqual(w.getMaximumSize(), [900, 600])
-    })
-  })
-
-  describe('BrowserWindow.setAspectRatio(ratio)', () => {
-    it('resets the behaviour when passing in 0', (done) => {
-      const size = [300, 400]
-      w.setAspectRatio(1 / 2)
-      w.setAspectRatio(0)
-      w.once('resize', () => {
-        expectBoundsEqual(w.getSize(), size)
-        done()
-      })
-      w.setSize(size[0], size[1])
-    })
-  })
-
-  describe('BrowserWindow.setPosition(x, y)', () => {
-    it('sets the window position', (done) => {
-      const pos = [10, 10]
-      w.once('move', () => {
-        const newPos = w.getPosition()
-        expect(newPos).to.deep.equal(pos)
-        done()
-      })
-      w.setPosition(pos[0], pos[1])
-    })
-  })
-
-  describe('BrowserWindow.setContentSize(width, height)', () => {
-    it('sets the content size', () => {
-      const size = [400, 400]
-      w.setContentSize(size[0], size[1])
-      const after = w.getContentSize()
-      expect(after).to.deep.equal(size)
-    })
-    it('works for a frameless window', () => {
-      w.destroy()
-      w = new BrowserWindow({
-        show: false,
-        frame: false,
-        width: 400,
-        height: 400
-      })
-      const size = [400, 400]
-      w.setContentSize(size[0], size[1])
-      const after = w.getContentSize()
-      expect(after).to.deep.equal(size)
-    })
-  })
-
-  describe('BrowserWindow.setContentBounds(bounds)', () => {
-    it('sets the content size and position', (done) => {
-      const bounds = { x: 10, y: 10, width: 250, height: 250 }
-      w.once('resize', () => {
-        expectBoundsEqual(w.getContentBounds(), bounds)
-        done()
-      })
-      w.setContentBounds(bounds)
-    })
-    it('works for a frameless window', (done) => {
-      w.destroy()
-      w = new BrowserWindow({
-        show: false,
-        frame: false,
-        width: 300,
-        height: 300
-      })
-      const bounds = { x: 10, y: 10, width: 250, height: 250 }
-      w.once('resize', () => {
-        expect(w.getContentBounds()).to.deep.equal(bounds)
-        done()
-      })
-      w.setContentBounds(bounds)
-    })
-  })
-
-  describe(`BrowserWindow.getNormalBounds()`, () => {
-    describe(`Normal state`, () => {
-      it(`checks normal bounds after resize`, (done) => {
-        const size = [300, 400]
-        w.once('resize', () => {
-          expectBoundsEqual(w.getNormalBounds(), w.getBounds())
-          done()
-        })
-        w.setSize(size[0], size[1])
-      })
-      it(`checks normal bounds after move`, (done) => {
-        const pos = [10, 10]
-        w.once('move', () => {
-          expectBoundsEqual(w.getNormalBounds(), w.getBounds())
-          done()
-        })
-        w.setPosition(pos[0], pos[1])
-      })
-    })
-    describe(`Maximized state`, () => {
-      before(function () {
-        if (isCI) {
-          this.skip()
-        }
-      })
-      it(`checks normal bounds when maximized`, (done) => {
-        const bounds = w.getBounds()
-        w.once('maximize', () => {
-          expectBoundsEqual(w.getNormalBounds(), bounds)
-          done()
-        })
-        w.show()
-        w.maximize()
-      })
-      it(`checks normal bounds when unmaximized`, (done) => {
-        const bounds = w.getBounds()
-        w.once('maximize', () => {
-          w.unmaximize()
-        })
-        w.once('unmaximize', () => {
-          expectBoundsEqual(w.getNormalBounds(), bounds)
-          done()
-        })
-        w.show()
-        w.maximize()
-      })
-    })
-    describe(`Minimized state`, () => {
-      before(function () {
-        if (isCI) {
-          this.skip()
-        }
-      })
-      it(`checks normal bounds when minimized`, (done) => {
-        const bounds = w.getBounds()
-        w.once('minimize', () => {
-          expectBoundsEqual(w.getNormalBounds(), bounds)
-          done()
-        })
-        w.show()
-        w.minimize()
-      })
-      it(`checks normal bounds when restored`, (done) => {
-        const bounds = w.getBounds()
-        w.once('minimize', () => {
-          w.restore()
-        })
-        w.once('restore', () => {
-          expectBoundsEqual(w.getNormalBounds(), bounds)
-          done()
-        })
-        w.show()
-        w.minimize()
-      })
-    })
-    describe(`Fullscreen state`, () => {
-      before(function () {
-        if (isCI) {
-          this.skip()
-        }
-        if (process.platform === 'darwin') {
-          this.skip()
-        }
-      })
-      it(`checks normal bounds when fullscreen'ed`, (done) => {
-        const bounds = w.getBounds()
-        w.once('enter-full-screen', () => {
-          expectBoundsEqual(w.getNormalBounds(), bounds)
-          done()
-        })
-        w.show()
-        w.setFullScreen(true)
-      })
-      it(`checks normal bounds when unfullscreen'ed`, (done) => {
-        const bounds = w.getBounds()
-        w.once('enter-full-screen', () => {
-          w.setFullScreen(false)
-        })
-        w.once('leave-full-screen', () => {
-          expectBoundsEqual(w.getNormalBounds(), bounds)
-          done()
-        })
-        w.show()
-        w.setFullScreen(true)
-      })
     })
   })
 
@@ -564,104 +261,6 @@ describe('BrowserWindow module', () => {
       it('is not available', () => {
         expect(w.setAutoHideCursor).to.be.undefined()
       })
-    })
-  })
-
-  describe('BrowserWindow.selectPreviousTab()', () => {
-    before(function () {
-      if (process.platform !== 'darwin') {
-        this.skip()
-      }
-    })
-
-    it('does not throw', () => {
-      expect(() => {
-        w.selectPreviousTab()
-      }).to.not.throw()
-    })
-  })
-
-  describe('BrowserWindow.selectNextTab()', () => {
-    before(function () {
-      if (process.platform !== 'darwin') {
-        this.skip()
-      }
-    })
-
-    it('does not throw', () => {
-      expect(() => {
-        w.selectNextTab()
-      }).to.not.throw()
-    })
-  })
-
-  describe('BrowserWindow.mergeAllWindows()', () => {
-    before(function () {
-      if (process.platform !== 'darwin') {
-        this.skip()
-      }
-    })
-
-    it('does not throw', () => {
-      expect(() => {
-        w.mergeAllWindows()
-      }).to.not.throw()
-    })
-  })
-
-  describe('BrowserWindow.moveTabToNewWindow()', () => {
-    before(function () {
-      if (process.platform !== 'darwin') {
-        this.skip()
-      }
-    })
-
-    it('does not throw', () => {
-      expect(() => {
-        w.moveTabToNewWindow()
-      }).to.not.throw()
-    })
-  })
-
-  describe('BrowserWindow.toggleTabBar()', () => {
-    before(function () {
-      if (process.platform !== 'darwin') {
-        this.skip()
-      }
-    })
-
-    it('does not throw', () => {
-      expect(() => {
-        w.toggleTabBar()
-      }).to.not.throw()
-    })
-  })
-
-  describe('BrowserWindow.addTabbedWindow()', () => {
-    before(function () {
-      if (process.platform !== 'darwin') {
-        this.skip()
-      }
-    })
-
-    it('does not throw', (done) => {
-      const tabbedWindow = new BrowserWindow({})
-      expect(() => {
-        w.addTabbedWindow(tabbedWindow)
-      }).to.not.throw()
-
-      expect(BrowserWindow.getAllWindows()).to.have.lengthOf(3) // Test window + w + tabbedWindow
-
-      closeWindow(tabbedWindow, { assertSingleWindow: false }).then(() => {
-        expect(BrowserWindow.getAllWindows()).to.have.lengthOf(2) // Test window + w
-        done()
-      })
-    })
-
-    it('throws when called on itself', () => {
-      expect(() => {
-        w.addTabbedWindow(w)
-      }).to.throw('AddTabbedWindow cannot be called by a window on itself.')
     })
   })
